requirements.txt
Telethon
cryptg
tgcrypto
pyrogram
python-decouple